client: transform store in reusable library
