+++ test/client/unit/specs/store/actions/tasks.js
@@ -1,24 +1,26 @@
-import { fetchUserTasks } from 'client/store/actions/tasks';
 import axios from 'axios';
-import storeModule from 'client/store';
-import cloneDeep from 'lodash/cloneDeep';
+import generateStore from 'client/store';
 
 describe('tasks actions', () => {
+  let store;
+
+  beforeEach(() => {
+    store = generateStore();
+  });
+
   describe('fetchUserTasks', () => {
     it('fetches user tasks', async () => {
-      const store = cloneDeep(storeModule);
       expect(store.state.tasks.loadingStatus).to.equal('NOT_LOADED');
       const tasks = [{_id: 1}];
       sandbox.stub(axios, 'get').withArgs('/api/v3/tasks/user').returns(Promise.resolve({data: {data: tasks}}));
 
-      await fetchUserTasks(store);
+      await store.dispatch('tasks:fetchUserTasks');
 
       expect(store.state.tasks.data).to.equal(tasks);
       expect(store.state.tasks.loadingStatus).to.equal('LOADED');
     });
 
     it('does not reload tasks by default', async () => {
-      const store = cloneDeep(storeModule);
       const originalTask = [{_id: 1}];
       store.state.tasks = {
         loadingStatus: 'LOADED',
@@ -28,14 +30,13 @@ describe('tasks actions', () => {
       const tasks = [{_id: 2}];
       sandbox.stub(axios, 'get').withArgs('/api/v3/tasks/user').returns(Promise.resolve({data: {data: tasks}}));
 
-      await fetchUserTasks(store);
+      await store.dispatch('tasks:fetchUserTasks');
 
       expect(store.state.tasks.data).to.equal(originalTask);
       expect(store.state.tasks.loadingStatus).to.equal('LOADED');
     });
 
     it('can reload tasks if forceLoad is true', async () => {
-      const store = cloneDeep(storeModule);
       store.state.tasks = {
         loadingStatus: 'LOADED',
         data: [{_id: 1}],
@@ -44,7 +45,7 @@ describe('tasks actions', () => {
       const tasks = [{_id: 2}];
       sandbox.stub(axios, 'get').withArgs('/api/v3/tasks/user').returns(Promise.resolve({data: {data: tasks}}));
 
-      await fetchUserTasks(store, true);
+      await store.dispatch('tasks:fetchUserTasks', true);
 
       expect(store.state.tasks.data).to.equal(tasks);
       expect(store.state.tasks.loadingStatus).to.equal('LOADED');
