+++ test/client/unit/specs/store/actions/user.js
@@ -1,24 +1,26 @@
-import { fetch as fetchUser } from 'client/store/actions/user';
 import axios from 'axios';
-import storeModule from 'client/store';
-import cloneDeep from 'lodash/cloneDeep';
+import generateStore from 'client/store';
 
-describe('user actions', () => {
+describe('tasks actions', () => {
+  let store;
+
+  beforeEach(() => {
+    store = generateStore();
+  });
+
   describe('fetch', () => {
     it('loads the user', async () => {
-      const store = cloneDeep(storeModule);
       expect(store.state.user.loadingStatus).to.equal('NOT_LOADED');
       const user = {_id: 1};
       sandbox.stub(axios, 'get').withArgs('/api/v3/user').returns(Promise.resolve({data: {data: user}}));
 
-      await fetchUser(store);
+      await store.dispatch('user:fetch');
 
       expect(store.state.user.data).to.equal(user);
       expect(store.state.user.loadingStatus).to.equal('LOADED');
     });
 
     it('does not reload user by default', async () => {
-      const store = cloneDeep(storeModule);
       const originalUser = {_id: 1};
       store.state.user = {
         loadingStatus: 'LOADED',
@@ -28,14 +30,13 @@ describe('user actions', () => {
       const user = {_id: 2};
       sandbox.stub(axios, 'get').withArgs('/api/v3/user').returns(Promise.resolve({data: {data: user}}));
 
-      await fetchUser(store);
+      await store.dispatch('user:fetch');
 
       expect(store.state.user.data).to.equal(originalUser);
       expect(store.state.user.loadingStatus).to.equal('LOADED');
     });
 
     it('can reload user if forceLoad is true', async () => {
-      const store = cloneDeep(storeModule);
       store.state.user = {
         loadingStatus: 'LOADED',
         data: {_id: 1},
@@ -44,7 +45,7 @@ describe('user actions', () => {
       const user = {_id: 2};
       sandbox.stub(axios, 'get').withArgs('/api/v3/user').returns(Promise.resolve({data: {data: user}}));
 
-      await fetchUser(store, true);
+      await store.dispatch('user:fetch', true);
 
       expect(store.state.user.data).to.equal(user);
       expect(store.state.user.loadingStatus).to.equal('LOADED');
