+++ website/client/components/appMenu.vue
@@ -130,7 +130,7 @@ $active-purple: #6133b4;
 </style>
 
 <script>
-import { mapState, mapGetters } from '../store';
+import { mapState, mapGetters } from 'client/libs/store';
 
 export default {
   computed: {
