+++ website/client/components/page.vue
@@ -4,7 +4,3 @@
     h2 Page
     p {{ $route.path }}
 </template>
-
-<script>
-export default { };
-</script>
