+++ website/client/components/parentPage.vue
@@ -3,7 +3,3 @@
   .col
     router-view
 </template>
-
-<script>
-export default { };
-</script>
