+++ website/client/store/getters/index.js
@@ -1,4 +1,4 @@
-import { flattenAndNamespace } from '../helpers/internals';
+import { flattenAndNamespace } from 'client/libs/store/helpers/internals';
 import * as user from './user';
 
 // Getters should be named as 'getterName' and can be accessed as 'namespace:getterName'
